refactor: split index.ts (657→165 lines) into focused modules
- state.ts (110 lines): all mutable state, load/save, group registration
- message-processor.ts (284 lines): processGroupMessages, agent output handling
- channel-manager.ts (177 lines): channel init, callbacks, onMessage/onChatMetadata
- index.ts (165 lines): main() startup sequence, signal handlers only

No behavior changes. All 450 tests pass. No circular dependencies.
