+++ src/state.ts
@@ -1,0 +1,110 @@
+// Copyright (c) 2026 Botler 360 SAS. All rights reserved.
+// See LICENSE.md for license terms.
+
+import fs from 'fs';
+import path from 'path';
+
+import {
+  getAllChats,
+  getAllRegisteredGroups,
+  getAllSessions,
+  getRouterState,
+  setRegisteredGroup,
+  setRouterState,
+} from './db.js';
+import { resolveGroupFolderPath } from './group-folder.js';
+import { logger } from './logger.js';
+import type { AvailableGroup } from './container-runner.js';
+import { RegisteredGroup } from './types.js';
+
+export let lastTimestamp = '';
+export let sessions: Record<string, string> = {};
+export let registeredGroups: Record<string, RegisteredGroup> = {};
+export let lastAgentTimestamp: Record<string, string> = {};
+
+// Track Google Chat reply targets: group folder -> gchat jid to reply to
+export const lastGchatReplyTarget: Record<string, string> = {};
+
+export function loadState(): void {
+  lastTimestamp = getRouterState('last_timestamp') || '';
+  const agentTs = getRouterState('last_agent_timestamp');
+  try {
+    lastAgentTimestamp = agentTs ? JSON.parse(agentTs) : {};
+  } catch {
+    logger.warn('Corrupted last_agent_timestamp in DB, resetting');
+    lastAgentTimestamp = {};
+  }
+  sessions = getAllSessions();
+  registeredGroups = getAllRegisteredGroups();
+  logger.info(
+    { groupCount: Object.keys(registeredGroups).length },
+    'State loaded',
+  );
+}
+
+export function saveState(): void {
+  setRouterState('last_timestamp', lastTimestamp);
+  setRouterState('last_agent_timestamp', JSON.stringify(lastAgentTimestamp));
+}
+
+export function registerGroup(jid: string, group: RegisteredGroup): void {
+  let groupDir: string;
+  try {
+    groupDir = resolveGroupFolderPath(group.folder);
+  } catch (err) {
+    logger.warn(
+      { jid, folder: group.folder, err },
+      'Rejecting group registration with invalid folder',
+    );
+    return;
+  }
+
+  registeredGroups[jid] = group;
+  setRegisteredGroup(jid, group);
+
+  // Create group folder
+  fs.mkdirSync(path.join(groupDir, 'logs'), { recursive: true });
+
+  logger.info(
+    { jid, name: group.name, folder: group.folder },
+    'Group registered',
+  );
+}
+
+/**
+ * Get available groups list for the agent.
+ * Returns groups ordered by most recent activity.
+ */
+export function getAvailableGroups(): AvailableGroup[] {
+  const chats = getAllChats();
+  const registeredJids = new Set(Object.keys(registeredGroups));
+
+  return chats
+    .filter((c) => c.jid !== '__group_sync__' && c.is_group)
+    .map((c) => ({
+      jid: c.jid,
+      name: c.name,
+      lastActivity: c.last_message_time,
+      isRegistered: registeredJids.has(c.jid),
+    }));
+}
+
+/** @internal - exported for testing */
+export function _setRegisteredGroups(
+  groups: Record<string, RegisteredGroup>,
+): void {
+  registeredGroups = groups;
+}
+
+// Mutable state setters for use by other modules
+export function setLastTimestamp(ts: string): void {
+  lastTimestamp = ts;
+}
+
+export function setLastAgentTimestamp(chatJid: string, ts: string): void {
+  lastAgentTimestamp[chatJid] = ts;
+}
+
+export function setSessionForGroup(folder: string, sessionId: string): void {
+  sessions[folder] = sessionId;
+}
